python/requirements.txt
supabase>=2.0.0
openpyxl>=3.1.2
python-dotenv>=1.0.1
typer[all]>=0.12.5